+++ src/FileSystem/FileSystem/SearchState.cs
@@ -27,6 +27,8 @@ internal class SearchState
 
     public DirectoryMatcher? DirectoryMatcher { get; }
 
+    public GlobMatcher? GlobFilter { get; set; }
+
     public Func<string, bool>? IncludeDirectory { get; set; }
 
     public Func<string, bool>? ExcludeDirectory { get; set; }
@@ -118,7 +120,7 @@ internal class SearchState
                     {
                         while (fi.MoveNext())
                         {
-                            FileMatch? match = MatchFile(fi.Current);
+                            FileMatch? match = MatchFile(fi.Current, directoryPath);
 
                             if (match is not null)
                                 yield return match;
@@ -150,7 +152,7 @@ internal class SearchState
                 if (matchStatus == MatchStatus.Success
                     || matchStatus == MatchStatus.Unknown)
                 {
-                    directoryMatch = MatchDirectory(directory.Path);
+                    directoryMatch = MatchDirectory(directory.Path, directoryPath);
 
                     if (directoryMatch is not null)
                     {
@@ -228,14 +230,44 @@ internal class SearchState
     {
         (FileMatch? fileMatch, Exception? exception) = FileMatcher!.Match(path, DefaultEncoding);
 
+        if (fileMatch is not null
+            && GlobFilter is not null)
+        {
+            string rootDirPath = Path.GetDirectoryName(path);
+
+            if (!GlobFilter.IsMatch(rootDirPath, path))
+                fileMatch = null;
+        }
+
         Report(path, SearchProgressKind.File, exception: exception);
 
         return fileMatch;
     }
 
-    public FileMatch? MatchDirectory(string path)
+    public FileMatch? MatchFile(string path, string rootDirectoryPath)
+    {
+        (FileMatch? fileMatch, Exception? exception) = FileMatcher!.Match(path, DefaultEncoding);
+
+        if (fileMatch is not null
+            && GlobFilter?.IsMatch(rootDirectoryPath, path) == false)
+        {
+            fileMatch = null;
+        }
+
+        Report(path, SearchProgressKind.File, exception: exception);
+
+        return fileMatch;
+    }
+
+    public FileMatch? MatchDirectory(string path, string rootDirectoryPath)
     {
         (FileMatch? fileMatch, Exception? exception) = DirectoryMatcher!.Match(path);
+
+        if (fileMatch is not null
+            && GlobFilter?.IsMatch(rootDirectoryPath, path) == false)
+        {
+            fileMatch = null;
+        }
 
         Report(path, SearchProgressKind.Directory, isDirectory: true, exception);